Improve code quality in various methods.
The parseCommand method is too long and should be refactored.

Splitting the behavior of this method into other methods and calling
these methods from parseCommand would allow the code to be more
readable.

Creating a method for commands that do not need to be specially
parsed and a method for those that do allows for greater separation
of responsibilities.